requirements.txt
google-api-python-client
google-auth-oauthlib 
google-auth-httplib2
google.generativeai
